It encrypts our user’s password and save it to db.
It encrypts our user’s password and save it to db. And if we are trying
to sing in the function, it takes the entered password and encrypts it
and checks it with the encrypted password which is saved in the db and
if they are same, it would log the user into the site.